move e2e test spec files to dedicated dir
