+++ testing/package.json
@@ -4,13 +4,12 @@
   "scripts": {
     "test:e2e": "npx playwright test --headed",
     "test:e2e:headless": "npx playwright test",
-    "playwright:test": "npx playwright test",
-    "playwright:open": "npx playwright test --ui",
+    "test:e2e:ui": "npx playwright test --ui",
     "test:downstream": "BASE_ADDRESS=http://localhost:9000 npx playwright test --grep @downstream",
     "test:downstream:remote": "npx playwright test --grep @downstream",
     "test:downstream:remote:docker": "bash run-e2e-docker.sh",
     "test:downstream:local": "BASE_ADDRESS=http://localhost:9000 npx playwright test --grep @downstream",
-    "test:github": "npx playwright test playwright/e2e/end-to-end-upstream.spec.ts"
+    "test:upstream": "npx playwright test playwright/e2e/upstream/"
   },
   "author": "",
   "license": "ISC",
