+++ testing/playwright/e2e/upstream/happy-path-upstream.spec.ts
@@ -1,13 +1,13 @@
 import { test } from '@playwright/test';
 
-import { setupCreatePlanIntercepts } from '../intercepts';
-import { CreatePlanWizardPage } from '../page-objects/CreatePlanWizard/CreatePlanWizardPage';
-import { PlanDetailsPage } from '../page-objects/PlanDetailsPage';
-import { PlansListPage } from '../page-objects/PlansListPage';
-import { createPlanTestData } from '../types/test-data';
+import { setupCreatePlanIntercepts } from '../../intercepts';
+import { CreatePlanWizardPage } from '../../page-objects/CreatePlanWizard/CreatePlanWizardPage';
+import { PlanDetailsPage } from '../../page-objects/PlanDetailsPage';
+import { PlansListPage } from '../../page-objects/PlansListPage';
+import { createPlanTestData } from '../../types/test-data';
 
 test.describe(
-  'Plans - Upstream End-to-End Migration',
+  'Plans - Upstream Happy Path Migration',
   {
     tag: '@upstream',
   },
